+++ src/sacct/process.c
@@ -44,7 +44,7 @@ char *find_hostname(uint32_t pos, char *hosts)
 	hostlist_t hostlist = NULL;
 	char *temp = NULL, *host = NULL;
 
-	if (!hosts || (pos == (uint32_t)NO_VAL))
+	if (!hosts || (pos == NO_VAL))
 		return NULL;
 
 	hostlist = hostlist_create(hosts);
