+++ src/scontrol/create_res.c
@@ -389,13 +389,13 @@ scontrol_create_res(int argc, char **argv)
 		goto SCONTROL_CREATE_RES_CLEANUP;
 	}
 	if (resv_msg.end_time == (time_t)NO_VAL &&
-	    resv_msg.duration == (uint32_t)NO_VAL) {
+	    resv_msg.duration == NO_VAL) {
 		exit_code = 1;
 		error("An end time or duration must be given.  No reservation created.");
 		goto SCONTROL_CREATE_RES_CLEANUP;
 	}
 	if (resv_msg.end_time != (time_t)NO_VAL &&
-	    resv_msg.duration != (uint32_t)NO_VAL &&
+	    resv_msg.duration != NO_VAL &&
 	    resv_msg.start_time + resv_msg.duration*60 != resv_msg.end_time) {
 		exit_code = 1;
 		error("StartTime + Duration does not equal EndTime.  No reservation created.");
