+++ src/slurmd/slurmd/req.c
@@ -3592,7 +3592,7 @@ _rpc_network_callerid(slurm_msg_t *msg)
 
 	uid_t req_uid = -1;
 	uid_t job_uid = -1;
-	uint32_t job_id = (uint32_t)NO_VAL;
+	uint32_t job_id = NO_VAL;
 	callerid_conn_t conn;
 	int rc = ESLURM_INVALID_JOB_ID;
 	char ip_src_str[INET6_ADDRSTRLEN];
@@ -4233,7 +4233,7 @@ _rpc_reattach_tasks(slurm_msg_t *msg)
 	int               fd;
 	uid_t             req_uid;
 	slurm_addr_t *cli = &msg->orig_addr;
-	uint32_t nodeid = (uint32_t)NO_VAL;
+	uint32_t nodeid = NO_VAL;
 	uid_t uid = -1;
 	uint16_t protocol_version;
